requirements.txt
itables==2.2.3
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pymongo==4.10.1
scipy==1.13.1